Adicionado service e mapper de pedidos
Implementado testes do mapper de pedidos de Dto para Entity

+++ src/main/java/br/com/jmsdevelopment/ecom/dto/pedido/ItemPedidoDto.java
@@ -1,5 +1,7 @@
 package br.com.jmsdevelopment.ecom.dto.pedido;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,5 +13,5 @@ public class ItemPedidoDto {
 	private Long idProduto;
 	private String nomeProduto;
 	private Integer quantidade;
-	private Double valorProduto;
+	private BigDecimal valorProduto;
 }
